Factor out compass dependency
I believe all the necessary parts that tie Lemonade to compass have been
removed. It should enable using compass features when available but
otherwise fall back to generating a plain url() object without compass.

There is now the ability to specify a sprites_dir where the source for
sprites are located. It contains the folders of single images that are
made into the sprite images. The combined images are saved in the
images_dir.

The images_dir defaults to 'public/images' or the Compass images_dir and
sprites_dir defaults to images_dir so behavior is the same as previous
versions.